Done reading of skin metadata and imported VersionInfo from trunk
git-svn-id: https://appetizer.googlecode.com/svn@405 ff61cf6c-2836-11df-8847-f3d5e5ef3dde

+++ branches/QtVersion/dialogs/ConfigDialog.cpp
@@ -71,7 +71,7 @@ void ConfigDialog::addSettingControlToLayout_(UserSetting* setting, QFormLayout*
     QComboBox* c = new QComboBox();
     control = static_cast<QComboBox*>(c);
 
-    for (int i = 0; i < setting->options().size(); i++) {
+    for (int i = 0; i < (int)setting->options().size(); i++) {
       std::pair<QString, QString> option = setting->options().at(i);
       c->addItem(option.first, QVariant(option.second));
     }
